requirements.txt
alabaster==0.7.9
autoenv==1.0.0
Babel==2.3.4
click==6.7
coverage==4.3.4
defusedxml==0.5.0rc1
Django==1.10.5
django-nose==1.4.4
djangorestframework==3.5.3
docutils==0.13.1
gitdb==0.6.4
GitPython==0.3.6
imagesize==0.7.1
Jinja2==2.9.5
MarkupSafe==0.23
nose==1.3.7
oauthlib==2.0.1
psycopg2==2.6.2
Pygments==2.2.0
PyJWT==1.4.2
pypi-publisher==0.0.4
python-dateutil==2.6.0
python-dotenv==0.6.2
python-openid==2.2.5
python3-openid==3.0.10
pytz==2016.10
requests==2.13.0
requests-oauthlib==0.7.0
git+git://github.com/AndelaOSP/shufflebox.git
smmap==0.9.0
snowballstemmer==1.2.1
social-auth-app-django==1.0.1
social-auth-core==1.1.0
Sphinx==1.5.2
sphinx-rtd-theme==0.1.9
wheel==0.24.0
